# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "kritzeLLM API",
    "summary": "Search and chat with a handwritten journal.",
...
    "version": "0.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Relative to wherever the app runs."
    }
  ],
  "paths": {
...
    "/blocks/{blockId}": {
      "get": {
        "tags": [
          "blocks"
        ],
        "summary": "Get a block",
        "description": "Returns a single block. Handy for resolving a chat source.",
        "operationId": "getBlock",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "blockId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^blk_[0-7][0-9a-hjkmnp-tv-z]{25}$",
              "description": "The block's ID.",
              "title": "Blockid"
            },
            "description": "The block's ID."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Block"
                }
              }
            }
          },
          "400": {
            "description": "The request couldn't be parsed (e.g. malformed JSON or multipart).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or wrong bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The resource doesn't exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Some values are invalid (including malformed IDs). Each one is listed in `errors`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "blocks"
        ],
        "summary": "Fix a block",
        "description": "Corrects a block's text, kind or `uncertain` flag. Its page is re-indexed in the background.",
        "operationId": "updateBlock",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "blockId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^blk_[0-7][0-9a-hjkmnp-tv-z]{25}$",
              "description": "The block's ID.",
              "title": "Blockid"
            },
            "description": "The block's ID."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockUpdateResult"
                }
              }
            }
          },
          "400": {
            "description": "The request couldn't be parsed (e.g. malformed JSON or multipart).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or wrong bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The resource doesn't exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Some values are invalid (including malformed IDs). Each one is listed in `errors`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/blocks/{blockId}/image": {
      "get": {
        "tags": [
          "blocks"
        ],
        "summary": "Get a block's image crop",
        "description": "Returns the part of the normalized page image that the block covers, plus a little padding.",
        "operationId": "getBlockImage",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "blockId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^blk_[0-7][0-9a-hjkmnp-tv-z]{25}$",
              "description": "The block's ID.",
              "title": "Blockid"
            },
            "description": "The block's ID."
          },
          {
            "name": "padding",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 0.25,
              "minimum": 0,
              "description": "Extra margin around the block, as a fraction of the page height.",
              "default": 0.02,
              "title": "Padding"
            },
            "description": "Extra margin around the block, as a fraction of the page height."
          }
        ],
        "responses": {
          "200": {
            "description": "A JPEG crop of the block.",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "The request couldn't be parsed (e.g. malformed JSON or multipart).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or wrong bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The resource doesn't exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Some values are invalid (including malformed IDs). Each one is listed in `errors`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List jobs",
        "description": "Lists background jobs, newest first.",
        "operationId": "listJobs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only jobs in this status.",
              "title": "Status"
            },
            "description": "Only jobs in this status."
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobKind"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only jobs of this kind.",
              "title": "Kind"
            },
            "description": "Only jobs of this kind."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of items to return.",
              "default": 25,
              "title": "Limit"
            },
            "description": "Maximum number of items to return."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "description": "The `nextCursor` from the previous page. Omit for the first page.",
              "title": "Cursor"
            },
            "description": "The `nextCursor` from the previous page. Omit for the first page."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobList"
                }
              }
            }
          },
          "400": {
            "description": "The request couldn't be parsed (e.g. malformed JSON or multipart).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or wrong bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Some values are invalid (including malformed IDs). Each one is listed in `errors`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{jobId}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get a job",
        "description": "Returns a job's status and progress. Poll this after a `202 Accepted`.",
        "operationId": "getJob",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^job_[0-7][0-9a-hjkmnp-tv-z]{25}$",
              "description": "The job's ID.",
              "title": "Jobid"
            },
            "description": "The job's ID."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "400": {
            "description": "The request couldn't be parsed (e.g. malformed JSON or multipart).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or wrong bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The resource doesn't exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Some values are invalid (including malformed IDs). Each one is listed in `errors`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{jobId}/result": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Download a job's result",
        "description": "Downloads the file a finished job produced (currently only exports).\n\nReturns `409` while the job hasn't succeeded.",
        "operationId": "getJobResult",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^job_[0-7][0-9a-hjkmnp-tv-z]{25}$",
              "description": "The job's ID.",
              "title": "Jobid"
            },
            "description": "The job's ID."
          }
        ],
        "responses": {
          "200": {
            "description": "The result file, with a `Content-Disposition` filename.",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "The request couldn't be parsed (e.g. malformed JSON or multipart).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or wrong bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The resource doesn't exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current state (e.g. a job is still running).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Some values are invalid (including malformed IDs). Each one is listed in `errors`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/admin/reindex": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Re-index everything",
        "description": "Rebuilds the search index for all pages, e.g. after switching the embedding model.\n\nTranscriptions are kept.",
        "operationId": "reindexAll",
        "responses": {
          "202": {
            "description": "Successful Response",
            "headers": {
              "Location": {
                "description": "URL of the job to poll.",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "400": {
            "description": "The request couldn't be parsed (e.g. malformed JSON or multipart).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or wrong bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current state (e.g. a job is still running).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "Block": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^blk_[0-7][0-9a-hjkmnp-tv-z]{25}$",
            "title": "Id",
            "description": "TypeID of a block."
          },
          "pageId": {
            "type": "string",
            "pattern": "^pg_[0-7][0-9a-hjkmnp-tv-z]{25}$",
            "title": "Pageid",
            "description": "TypeID of a page."
          },
          "seq": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Seq",
            "description": "Position in reading order, starting at 0."
          },
          "kind": {
            "$ref": "#/components/schemas/BlockKind"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "The transcribed text. Unreadable words are written as `[?]`."
          },
          "uncertain": {
            "type": "boolean",
            "title": "Uncertain",
            "description": "`true` when the model wasn't sure about parts of this block."
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          }
        },
        "type": "object",
        "required": [
          "id",
          "pageId",
          "seq",
          "kind",
          "text",
          "uncertain",
          "location"
        ],
        "title": "Block",
        "description": "A located piece of text on a page."
      },
...
      "BlockKind": {
        "type": "string",
        "enum": [
          "heading",
          "paragraph",
          "list",
          "marginNote",
          "drawing",
          "other"
        ],
        "title": "BlockKind",
        "description": "What kind of thing a block is on the page. Drawings are described in words, in brackets."
      },
      "BlockUpdate": {
        "properties": {
          "kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BlockKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Text",
            "description": "Corrected text."
          },
          "uncertain": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uncertain",
            "description": "Whether parts of the text are still unsure."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "BlockUpdate",
        "description": "Corrections to a single block. Only the given fields change."
      },
      "BlockUpdateResult": {
        "properties": {
          "block": {
            "$ref": "#/components/schemas/Block"
          },
          "job": {
            "$ref": "#/components/schemas/Job"
          }
        },
        "type": "object",
        "required": [
          "block",
          "job"
        ],
        "title": "BlockUpdateResult",
        "description": "The corrected block, plus the re-index job."
      },
...
      "Job": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^job_[0-7][0-9a-hjkmnp-tv-z]{25}$",
            "title": "Id",
            "description": "TypeID of a job."
          },
          "kind": {
            "$ref": "#/components/schemas/JobKind"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "target": {
            "$ref": "#/components/schemas/JobTarget"
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobProgress"
              },
              {
                "type": "null"
              }
            ],
            "description": "How far along the job is, when it can tell."
          },
          "resultUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resulturl",
            "description": "Where to download the result once `succeeded` (exports only)."
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Problem"
              },
              {
                "type": "null"
              }
            ],
            "description": "Why the job failed; `null` unless `status` is `failed`."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat",
            "description": "RFC 3339 timestamp in UTC."
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "description": "RFC 3339 timestamp in UTC."
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat",
            "description": "When a worker picked the job up."
          },
          "finishedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "description": "RFC 3339 timestamp in UTC."
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat",
            "description": "When the job succeeded or failed."
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "status",
          "target",
          "progress",
          "resultUrl",
          "error",
          "createdAt",
          "startedAt",
          "finishedAt"
        ],
        "title": "Job",
        "description": "A unit of background work. Poll it until `status` is `succeeded` or `failed`."
      },
      "JobKind": {
        "type": "string",
        "enum": [
          "ingestPage",
          "transcribePage",
          "reindexPage",
          "reindexAll",
          "exportJournal"
        ],
        "title": "JobKind",
        "description": "What a job does."
      },
      "JobList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Job"
            },
            "type": "array",
            "title": "Items"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor",
            "description": "Pass as `cursor` to get the next page; `null` on the last page."
          }
        },
        "type": "object",
        "required": [
          "items",
          "nextCursor"
        ],
        "title": "JobList",
        "description": "A page of jobs."
      },
      "JobProgress": {
        "properties": {
          "done": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Done",
            "description": "Units finished so far."
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "description": "Total units of work."
          }
        },
        "type": "object",
        "required": [
          "done",
          "total"
        ],
        "title": "JobProgress",
        "description": "How far along a job is."
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed"
        ],
        "title": "JobStatus",
        "description": "Lifecycle of a job."
      },
      "JobTarget": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/JobTargetType"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "ID of the page or journal; `null` for `all`."
          }
        },
        "type": "object",
        "required": [
          "type",
          "id"
        ],
        "title": "JobTarget",
        "description": "What the job works on."
      },
      "JobTargetType": {
        "type": "string",
        "enum": [
          "page",
          "journal",
          "all"
        ],
        "title": "JobTargetType",
        "description": "Kind of thing a job works on."
      },
...
      "Location": {
        "properties": {
          "box": {
            "$ref": "#/components/schemas/Box"
          },
          "bands": {
            "$ref": "#/components/schemas/Bands"
          },
          "side": {
            "$ref": "#/components/schemas/Side"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "description": "Human-friendly description of the spot, e.g. \"middle of the page, paragraph 3\"."
          }
        },
        "type": "object",
        "required": [
          "box",
          "bands",
          "side",
          "label"
        ],
        "title": "Location",
        "description": "Where a block sits on its page."
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Identifies the problem type; `about:blank` when there's nothing more specific.",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Short, human-readable summary of the problem type."
          },
          "status": {
            "type": "integer",
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "description": "The HTTP status code."
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "What went wrong in this particular case."
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance",
            "description": "The request path that caused the problem."
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProblemError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Individual validation problems (only for `422`)."
          }
        },
        "type": "object",
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "description": "An RFC 9457 problem details object."
      },
      "ProblemError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "What's wrong with this value."
          },
          "pointer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pointer",
            "description": "JSON Pointer to the offending value in the request body, e.g. `/name`."
          },
          "parameter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parameter",
            "description": "Name of the offending query, path or header parameter."
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ProblemError",
        "description": "One invalid value in a request."
      },
...
      "Side": {
        "type": "string",
        "enum": [
          "left",
          "right",
          "full"
        ],
        "title": "Side",
        "description": "Horizontal placement. `full` spans the text width; `left`/`right` are margin notes or columns."
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "The `API_TOKEN` configured on the server. Only needed when one is set.",
        "scheme": "bearer"
      }
    }
  },
...
}